+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v1.1.14
-* 
+* One-line requirement no longer needed to use RMT.
+    * Class references are automatically added if a mod has __initialize in it.
+* __initialize will now print a message to console if a mod's __initialize/__post_initialize fails to run.
 
 ### v1.1.13
 * Instance : get_data tables and Instance callbacks are no longer instantly deleted for players on-death.